+++ backend/src/message/message.module.ts
@@ -1,11 +1,11 @@
 import { Module } from "@nestjs/common";
 import { ConversationService } from "./conversation.service";
-import { MessageGateway } from "./message.gateway";
 import { AuthModule } from "../auth/auth.module";
 import { SequelizeModule } from "@nestjs/sequelize";
 import { MessageModel } from "./models/message.model";
 import { ActiveConversationModel } from "./models/active_conversation.model";
 import { ConversationModel } from "./models/conversation.model";
+import { MessageGateway } from "./message.gateway";
 
 @Module({
   imports: [