+++ backend/src/user/models/user.model.ts
@@ -7,6 +7,7 @@ import { PostModel } from "../../post/models/post.model";
 import { PhotoModel } from "./photo.model";
 import { ConversationModel } from "../../message/models/conversation.model";
 import { MessageModel } from "../../message/models/message.model";
+import { FollowModel } from "./follow.model";
 
 interface UserCreationAttr {
   email: string;
@@ -94,21 +95,25 @@ export class UserModel extends Model<UserModel, UserCreationAttr> {
   @HasMany(() => BlockedUserModel)
   blockedUsers: BlockedUserModel[];
 
-  @ApiProperty({ example: [MessageModel], description: "User's roles" })
+  @ApiProperty({ example: [RoleModel], description: "User's roles" })
   @BelongsToMany(() => RoleModel, () => UserRolesModel)
   roles: RoleModel[];
 
-  @ApiProperty({ example: [MessageModel], description: "User's posts" })
+  @ApiProperty({ example: [PostModel], description: "User's posts" })
   @HasMany(() => PostModel)
   posts: PostModel[];
 
-  @ApiProperty({ example: [MessageModel], description: "User's photos" })
+  @ApiProperty({ example: [PhotoModel], description: "User's photos" })
   @HasMany(() => PhotoModel)
   photos: PhotoModel[];
 
-  @ApiProperty({ example: [MessageModel], description: "User's conversations" })
-  @HasMany(() => ConversationModel)
+  @ApiProperty({ example: [ConversationModel], description: "User's conversations" })
+  @HasMany(() => ConversationModel, "firstUser" && "secondUser")
   conversations: ConversationModel[];
+
+  @ApiProperty({ example: [FollowModel], description: "User's follows" })
+  @HasMany(() => FollowModel)
+  follows: FollowModel[];
 
   @ApiProperty({ example: [MessageModel], description: "User's messages" })
   @HasMany(() => MessageModel)