+++ frontend/src/pages/Messages/Conversation.tsx
@@ -1,10 +1,12 @@
 import React from "react";
-import { Grid, Typography } from "@mui/material";
 import { Messages } from "./Messages";
 import { AddMessageForm } from "./AddMessage";
-import { Socket } from "socket.io-client";
 import { useParams } from "react-router-dom";
 import { MessageDto } from "../../types/message/message.dto";
+import { Grid, Typography } from '@mui/material';
+import { Socket } from 'socket.io-client';
+// @ts-ignore
+import styles from './MessagePage.module.css';
 
 interface IConversation {
 	socket: Socket | null;
@@ -20,22 +22,24 @@ export const Conversation: React.FC<IConversation> = ({ socket, userId }) => {
 		if (!params.id) {
 			return;
 		}
-		socket?.emit('joinConversation', {friendId: Number(params.id)});
+		socket?.emit('joinConversation', {conversationId: Number(params.id)});
 		socket?.on("messages", (data) => {
-			console.log(data);
-			setMessages([...messages, ...data]);
+			setMessages((messages) => [...messages, ...data])
 		});
-	},[params])
+		socket?.on("newMessage", (data) => {
+			setMessages((messages) => [...messages, data])
+		});
+	},[params, socket, params.id])
 
 	return (
-		<Grid>
+		<Grid className={styles.conversation}>
 			{!params.id ? <Grid>
 				<Typography>Start chat!</Typography>
 			</Grid>
 			:
 			<Grid>
-				<Messages socket={socket} messages={messages}/>
-				<AddMessageForm socket={socket} userId={userId} />
+				<Messages socket={socket} messages={messages} userId={userId}/>
+				<AddMessageForm socket={socket} userId={userId} conversationId={Number(params.id)} />
 			</Grid>
 			}