+++ frontend/src/pages/Messages/Message.tsx
@@ -3,14 +3,15 @@ import { MessageDto } from "../../types/message/message.dto";
 import { Grid } from "@mui/material";
 
 interface MessageInterface {
-	message: MessageDto
+	message: MessageDto;
+	userId: number | null;
 }
 
-export const Message:React.FC<MessageInterface> = ({ message }) => {
+export const Message:React.FC<MessageInterface> = ({ message, userId }) => {
 
 	return (
 		<Grid>
-			{message.message}
+			{message.senderId === userId && "Me: "}{message.text}
 		</Grid>
 	);
 };