+++ frontend/src/pages/Messages/Messages.tsx
@@ -6,14 +6,15 @@ import { Grid } from "@mui/material";
 
 
 interface IMessage {
-	socket: Socket | null,
-	messages: MessageDto[]
+	socket: Socket | null;
+	messages: MessageDto[];
+	userId: number | null;
 }
 
-export const Messages: React.FC<IMessage> = ({ socket, messages  }) => {
+export const Messages: React.FC<IMessage> = ({ socket, messages, userId  }) => {
 
 
 	return <Grid style={{height: '400px', overflowY: 'auto'}}>
-		{messages.map((m, index) => <Message key={index} message={m}/>)}
+		{messages.map((m, index) => <Message key={index} message={m} userId={userId}/>)}
 	</Grid>
 }