+++ frontend/src/pages/Profile/ProfileGallery/ProfileGallery.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from 'react-router-dom';
 import styles from "./ProfileGallery.module.css";
 
 interface ProfileGalleryInterface {
-	images: PhotoDto[];
+	images?: PhotoDto[];
 	userId: number;
 	gallery: boolean;
 	setGallery: (value: boolean) => void;
@@ -29,7 +29,7 @@ export const ProfileGallery: React.FC<ProfileGalleryInterface> = ({ images, user
 				<Typography variant="h6" align="center">Gallery</Typography>
 				<Button variant="contained" onClick={onClick}>Back</Button>
 				<Grid className={styles.galleryGrid}>
-					{images.map((image) => {
+					{images?.map((image) => {
 						return <>
 							<img className={styles.galleryImage}  src={"http://localhost:8000/" + image.url} alt="" />
 						</>;