+++ frontend/src/types/profile/follow.dto.ts
@@ -1,5 +1,4 @@
 export interface FollowDto {
     userId: number;
     followId: number;
-
 }